remove handlers from provider and introduce new hook
