+++ deployables/extension/src/panel/providers-ui/index.ts
@@ -1,1 +1,2 @@
 export { ProvideProvider, useProvider } from './ProvideProvider'
+export { useSendTransaction } from './useSendTransaction'
